+++ src/api/axios.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosError, AxiosRequestConfig } from 'axios';
-import { getToken, removeToken, setToken } from '@/utils/tokenUtils';
+import { getToken, removeToken } from '@/utils/tokenUtils';
 
 import { refreshToken } from '@/components/login/queries';
 import { toast } from 'sonner';
@@ -19,62 +19,6 @@ const handleTokenError = (message: string) => {
   window.location.href = '/login';
 };
 
-const parseJwt = (token: string) => {
-  try {
-    const base64Url = token.split('.')[1];
-    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-    const jsonPayload = decodeURIComponent(
-      atob(base64)
-        .split('')
-        .map(function (c) {
-          return `%${`00${c.charCodeAt(0).toString(16)}`.slice(-2)}`;
-        })
-        .join(''),
-    );
-
-    return JSON.parse(jsonPayload);
-  } catch (error) {
-    return null;
-  }
-};
-
-const getAccessTokenExpiration = () => {
-  const token = getToken();
-  if (token) {
-    const parsedToken = parseJwt(token);
-    if (parsedToken && parsedToken.exp) {
-      return parsedToken.exp * 1000;
-    }
-  }
-  return null;
-};
-
-const tokenRefresh = () => {
-  const expiration = getAccessTokenExpiration();
-  if (expiration) {
-    const currentTime = Date.now();
-    const timeout = expiration - currentTime - 60000;
-
-    if (timeout > 0) {
-      setTimeout(async () => {
-        try {
-          const refreshData = await refreshToken();
-          const newAccessToken = getToken();
-          if (!newAccessToken) {
-            throw new Error('리프레시 토큰이 만료되었습니다.');
-          }
-          setToken(newAccessToken);
-          tokenRefresh();
-        } catch (error) {
-          handleTokenError(
-            '토큰 갱신 중 오류 발생되었습니다. 다시 로그인해주세요.',
-          );
-        }
-      }, timeout);
-    }
-  }
-};
-
 export const createClient = (config?: AxiosRequestConfig) => {
   const axiosInstance = axios.create({
     baseURL: import.meta.env.VITE_API_URL,
@@ -87,23 +31,8 @@ export const createClient = (config?: AxiosRequestConfig) => {
   });
 
   axiosInstance.interceptors.request.use(async (request) => {
-    let accessToken = getToken();
-
-    if (!accessToken) {
-      try {
-        const refreshData = await refreshToken();
-
-        accessToken = getToken();
-        if (accessToken) {
-          request.headers.Authorization = `Bearer ${accessToken}`;
-        } else {
-          throw new Error('리프레시 토큰이 만료되었습니다.');
-        }
-      } catch (error) {
-        handleTokenError('토큰이 만료되었습니다. 다시 로그인해주세요.');
-        return Promise.reject(error);
-      }
-    } else {
+    const accessToken = getToken();
+    if (accessToken) {
       request.headers.Authorization = `Bearer ${accessToken}`;
     }
     return request;
@@ -124,7 +53,7 @@ export const createClient = (config?: AxiosRequestConfig) => {
 
         if (errorMessage === '토큰이 만료되었습니다.') {
           try {
-            await refreshToken();
+            refreshToken();
             const newAccessToken = getToken();
             if (!newAccessToken)
               throw new Error('리프레시 토큰이 만료되었습니다.');
@@ -169,4 +98,3 @@ export const createClient = (config?: AxiosRequestConfig) => {
 };
 
 export const httpClient = createClient();
-tokenRefresh();
