+++ src/components/home/queries.ts
@@ -2,14 +2,7 @@ import { API_END_POINT } from '@/constants/api';
 import { httpClient } from '@/api/axios';
 import { queryKeys } from '@/constants/queryKeys';
 import type { PreRegisterAuction, RegisterAuction } from 'Auction';
-import {
-  UseMutationResult,
-  UseQueryResult,
-  useMutation,
-  useSuspenseQueries,
-} from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
-import { refreshToken } from '../login/queries';
+import { useSuspenseQueries } from '@tanstack/react-query';
 
 export const useGetHomeAuctions = () => {
   const getBestAuctions = async (): Promise<RegisterAuction[]> => {
@@ -45,28 +38,4 @@ export const useGetHomeAuctions = () => {
   const preRegisterAuctions = preRegisterAuctionsQuery.data;
 
   return { bestAuctions, imminentAuctions, preRegisterAuctions };
-};
-
-export const useRefreshTokenOnSuccess = (): UseMutationResult<
-  void,
-  unknown,
-  void,
-  unknown
-> => {
-  const navigate = useNavigate();
-  const queryParams = new URLSearchParams(window.location.search);
-  const status = queryParams.get('status');
-
-  const RefreshMutation = useMutation({
-    mutationFn: () => refreshToken(),
-    onSuccess: () => {
-      navigate('/');
-    },
-    onError: (error) => {},
-  });
-  if (status === 'success') {
-    RefreshMutation.mutate();
-  }
-
-  return RefreshMutation;
 };
